Modify UploadContactsService to UploadContactsJob, upgrade lib 'libphonenumber'
